reorganize tests into a single directory
