+++ tests/Makefile
@@ -1,0 +1,6 @@
+all:
+	( cd functional && make )
+	( cd singlestep && make )
+	@echo ===============================
+	@echo === ALL TESTS COMPLETE: success
+	@echo ===============================
