+++ tests/functional/Makefile
@@ -15,7 +15,9 @@ SHELL := /bin/bash
 BASE := 6502_65C02_functional_tests
 
 all: $(BASE) $(BASE)/as65_142 main tests tests
-   @echo TEST COMPLETE: success
+	@echo ======================================
+	@echo === FUNCTIONAL TESTS COMPLETE: success
+	@echo ======================================
 
 clean:
 	rm -rf $(BASE) ft.* dt.* it.*
@@ -30,8 +32,8 @@ $(BASE)/as65_142:
 	mkdir -p $(BASE)/as65_142
 	( cd $(BASE)/as65_142 && unzip ../as65_142.zip )
 
-main: main.cpp ../mos6502.cpp ../mos6502.h
-	g++ -Wall -O3 -o main ../mos6502.cpp main.cpp
+main: main.cpp ../../mos6502.cpp ../../mos6502.h
+	g++ -Wall -O3 -o main ../../mos6502.cpp main.cpp
 
 tests: 6502_functional_test 6502_decimal_test 6502_interrupt_test
