+++ tests/functional/as65.sh
