+++ tests/functional/decimal.patch
