+++ tests/functional/main.cpp
@@ -1,6 +1,6 @@
-// compile with "g++ main.cpp ../mos6502.cpp -o main"
+// compile with "g++ main.cpp ../../mos6502.cpp -o main"
 
-#include "../mos6502.h"
+#include "../../mos6502.h"
 
 #include <stdlib.h>
 #include <string.h>
